refactor(tap-net): split the crate root into tap, allocation, and tests
The moved `lib.rs` was 1268 lines. The TAP device plumbing (create over
`/dev/net/tun`, bring-up, point-to-point addressing, destroy) becomes the
Linux-only `tap` module as free functions — it never read manager state;
`NetworkAllocation` and its netmask tests become `allocation`; the
manager's unit tests move to `tests.rs`. No behavior change; the two
root-detection helpers pick up the `is_ok_and`/`is_some_and` forms clippy
asks for on the Linux target.
